[typ] Only create pools when necessary
Updates typ's pool creation logic to remove cases where we can create
a pool when it is unnecessary. The main cases for this are:

1. We know all tests will be run serially due to --jobs=1, in which case
   we do not need to bother with a parallel pool.
2. We know a serial/parallel pool won't be used because we don't have any
   relevant tests.

In practice, this only saves a few seconds per test invocation. However,
this can add up if a test is heavily sharded, as each swarming task will
get those savings.

Bug: chromium:1363409
Change-Id: I528232c91b84d35e8946a69935e68b2b7fb5a52e