[chromeperf cloud functions] update cabe grpc stubs
These were atutomatically gernerated by running `make generate` in the
functions/ directory, with the skia/buildbot repo checked out as a
sibling to the chromium/src checkout on my workstation.

The updates are necessary to keep up with updates to cabe's grpc API
and the protocol buffer message types it uses.

This update was run using this .proto update CL in the skia/buildbot
repo: https://skia-review.googlesource.com/c/buildbot/+/705984

Change-Id: I4b7209b28823b2c780d3a8c56b5c3bf7f8135948