[heap] Clean up obsolete blink-gc metrics
This CL removes the obsolete blink-gc-* metrics, which aggregated over
the obsolete BlinkGC.* events. They are now part of the v8:gc:*:cpp metrics. The CL removes their calculation and tests. It also introduces
v8-gc-minor-mark-compactor, corresponding to the V8.GCMinorMC event,
analogously to v8-gc-scavenger which corresponds to the V8.GCScavenger
event.

Bug: chromium:1154636
Change-Id: I9925ad6c1d02d1444096fbb6b12bb2ecf1828d56