[typ] Fix expected results edge case
Fixes an edge case in typ's Runner._run_one_test when the act of
running a test causes expected results to change. The most common known
case of this is for GPU tests that restart the browser with specific
browser arguments. The change in browser args can cause the generated
tags to be updated, which in turn can affect which expectations apply
to the test.

Bug: chromium:1454250
Change-Id: Ifd63560576617468af8e657ad1b4c4754c6c07b4