[chromeperf] fix failing cloud build for alert group test
Cloud build is failing on test testCreatingUngrouped. The new_groups are not assigned before referred to because exception was thrown when trying to publish to cloud metric in the parity logic.

As the parity logic will be removed and it is now wrapped in a try block, I'll allow the exception here and move the assignment of new_groups up so that the exception will not interrupt it.

Bug: chromium:1439636
Change-Id: I720f5ad954401044a13c2ad4eb2160599422bf18