Support filtering httparchive requests by response status code
Use case: Most of the time we don't want status 500 responses in the
final archive.

In https://crbug.com/1466157 we accidentally recorded a error 500
request on the initial page load which cases the benchmark to behave
inconsistently between running from archive vs the live page.

Currently the benchmark has to navigate twice to get past the original
failing request, which is not ideal. By adding filtering by status
code to httparchive we can easily fix this post-recording.

Bug: chromium:1466157
Change-Id: I3fc86f8da352e5237032eefc20c81d21ad408b03