[sandwich] set analysis_spec to limit what cabe needs to analyze
sandwich verification for bisect jobs should only need to analyze
the "chart" (aka "measurement" or "workload") that was implicated in
the regression, and can safely ignore any other data reported by
the pinpoint jobs.

Without this change, the sandwich verification workflow is
asking cabe to analyze *everything* produced by the A/B job, including
values that break some assumptions required by the wilcoxon test,
which in turn causes the entire GetAnalysisRequest to fail.

With this change, the worfklow handler provides a more specific
scope of analysis that lets cabe ignore the problem data that was causing it to fail.

In the process of fixing this, I had to regenerate some python
bindings for the cabe grpc protos. The _pb2.py and _pb2_grpc.py
files are automatically generated, so I don't recommend trying
to review them in detail.

Bug: chromium:1469724
Change-Id: Idc55c170004c32724a8660256c0c7313fd454b09