[sandwich] log call steps for every execution
Without this logging, all we get in the GCP console UI are log lines for
start and end events for the overall execution.

Logging all of the calls will add lines for each workflow
step the calls something. e.g. postStartPinpointJob will log the call
result, including the pinpoint job ID.  Currently this ID is hard to
track down because you have to browse the pinpoint main page looking
for jobs that might be the one started by your workflow.

Change-Id: I1c27cf6c39b2a153942fa455281933d8e23802aa