Access the cache copy of install_attributes, not disk copy
As per install_attributes doc, "All the daemons that need the
information are supposed to use only the file in memory."

This source only checks the existence of this install_attributes
file. This is a good reason to rather access the cache copy.

For context, we are moving out install_attributes related features
from cryptohome to a new service. install_attributes.pb file will
no longer be saved in /home/.shadow/ dir once the migration is
complete. But the path of the identical cache copy will remain
unchanged. Relevant bug: http:// b/292043969

Change-Id: I11bc39054d3a91bda7983a34332075068e730d92