[pinpoint] add comments to unused graph, task and evaluator
Graphs, tasks and evaluators were all introduced in an attempt to
integrate with Swarming's PubSub, but the code was left in
experimental mode. This change adds comments with references to
crbug/988912 and crrev.com/c/1861616 for context and a note stating
that the section of code in unused in production.

Bug: chromium:1471259
Change-Id: If32439775d2f3c4f34ac0250dfb461a20028f9ef