Fix TsProxyServer restart
Fixes TsProxyServer being erroneously permanently shut down if an error
occurs during a restart. Previously, if an error occurred when bringing
the server back up, the server would remain dead indefinitely since it
looked like the server was not supposed to be running. Now, we keep
track of whether the last open attempt was successful and use that to
determine whether the server is intentionally closed or not.

Bug: dawn:2024
Change-Id: I90f01fec4a04334bf17f1f2ba108bc5cb18e0d43