Strip `AND_CHMOD` from unzipping script.
This change is required to fix a problem where we push 3GB of data
onto test devices.

We currently use a small Java `Unzip` utility to extract zip file
contents on Android.

The tool does not apply any permissions to files it decompresses,
meaning it follows whatever `umask` mandates.

The `/sdcard` partition does not allow
- supplying executable bit to files,
- stripping executable bit from folders,
- giving any permission to all users (only user and group permissions
  are accepted).

meaning the `chmod` part is irrelevant.

The challenge posed by this script is it is susceptible to exceeding
maximum permitted cmdline length.

Bug: 1478271
Change-Id: Ia83b253d8aae2e60410f61cf38b116db543de326