[sandwich] switch from allowlisting to blocklisting
This should open up sandwich verification to a much larger set of
regressions.

The configs included in BLOCKED_DEVICES are chosen based on the number
of devices we have available for running pinpoint A/B jobs.  If we
don't have enough devices, we shouldn't try to run sandwich verification
on them.

Change-Id: I03ca240077a0dd6e1178c76f98b088dcd586adbd