[heap] Force GCs on memory dump for all renderers
system_health.memory_desktop benchmarks (and others) collect memory
dumps from all processes. To get an accurate view of the renderer's
live memory, it is important to GC the renderer before getting the
dump. The benchmark therefore force a GC on the main renderer.
HOwever, in some (most?) stories there are multiple renderers and only
one of them gets GCed. The rest are thus reporting memory statistics
of dead objects.
Usually, v8's new space is empty after a full GC. As a symptom of not doing a GC before the dump, we were observing non-empty new spaces.
Piping the `deterministic_mode` from `action_runner.MeasureMemory`
all the way done to the cpp code implementing the memory dump resolves
the issue and forces a GC for all renderers.

Bug: v8:12612
Change-Id: Id49a01ba190a5c7a17e96021d634bcf2abc7cdb3