[sandwich] switch from allowlisting to blocklisting
This should open up sandwich verification to a much larger set of
regressions.

After this lands, any regression reported with a combination of subscription, benchmark or device that are all NOT explicitly blocked will be allowed into the sandwich verification process.

The configs included in BLOCKED_DEVICES are chosen based on the number of devices we have available for running pinpoint A/B jobs.  If we don't have enough devices, we shouldn't try to run sandwich verification on them.

BLOCKED_SUBSCRIPTIONS and BLOCKED_BENCHMARKS are currently empty, but if we discover there are specific problem subscriptions or benchmarks in prod we can add them here.


Change-Id: I03ca240077a0dd6e1178c76f98b088dcd586adbd