[Chromeperf] Reduce pinpoint cycle time
The existing pinpoint scheduler schedules new execution and poll the status, but does not further schedule new execution if the current one is done. This will delay the pinpoint for another one minute even if the status is ready, e.g., when the isolate is built by waterfall and ready for use by pinpoint.
This CL switches the order of poll and schedule, and use a loop to keep exploring finished execution.

Change-Id: I5c958fae94c986a8a71c4d245c40f8d8d13e9035