[chromeperf] fix broken alert_groups_test with mock workflow_service
unit tests are failing on cloud build because this test is actually
calling the real cloud workflows service(!). Example failure:
https://pantheon.corp.google.com/cloud-build/builds;region=global/7bc26599-68c6-4554-8fe4-cb0c6056bd36;step=2?e=-13802955&mods=component_inspector&project=chromeperf

Note that this is probably passing on dev workstations because the
tests are likely running with the dev's gcloud auth credentials set.

No idea how this has been passing on our CQ testers though.

Change-Id: I4ab9e3d515f60c6ec459c3f29c63a82474be6078