------------------------------------------------------------------------ Fix testPressEndAndHome
Browser uses Google API keys to get access to Google services.
If browser was built without Google API keys, then on start browser
would try to fetch Google API keys from env variables. If those variables
are not set a warning "Google API Keys Missing" will appear at the start.
This warning disappears if any action is performed in the browser window.
This behaviour causes flaky bug in tests. The warning occupies part
of the window space and because of this method "_window_height" returns
different values when the warning is on the screen and
when it disappeared. It is critical for tests with scroll position checks
like testPressEndAndHome.
Tests in KeyPressActionTest class do not use any Google APIs,
so it is not needed to provide Google API keys to browser
to run those tests. If browser was built with Google API keys,
the warning would not be shown on start, so it is no need to do anything.
Otherwise we need to set specific env variables to 'no'
to disable warning.

Change-Id: I621756098ea1a3d093daa6caae3e94225cf55bbe