[chromeperf] ignore "anomalies" measured by different physical devices
See the bug for more context but tl;dr this CL aggressively filters out anomalies that span swarming device changes from being processed as alerts that humans need to look into.

I got hit with device-change "anomalies" on this past week's on-call rotation, and I got hit with it on my previoew rotation too.

A recurring source of pain is when, for ...whatever reason...
a device that was reporting measurements for some metric on the
waterfall is replaced by an entirely different physical device who
then proceeds to report that same metric but with a different bias
than the previous device that reported it.

You end up with a "change" in the timeseries values according to chromeperf, but what it actually reflects is differences in *physical device* performance, not differences in the *code under test*'s performance.

This CL:

- adds bot_id to models.graph_data.Row as an indexed field. Row.bot_id exists in a sense already because Row is an ndb.Expando, and we have code already that reads and writes the field, but it's not currently indexed and won't show up in ndb queries that use projections.

- adds bot_id_before_anomaly and bot_id_after_anomaly to models.anomaly.Anomaly to keep track of which devices reported the measurements on either side of the Anomaly

- updates find_anomalies.py to populate the bot_id_before_anomaly and bot_id_after_anomaly fields when it creates new Anomaly entities

- updates alert_group_workflow.py to ignore Anomalies from GroupUpdates where the before and after bot_ids on the Anomaly don't match.

Bug: chromium:1485723
Change-Id: I34347f1e67a7cb2159cec05c4b87fd6ed115730c