[chromeperf] ignore "anomalies" measured by different physical devices
See the bug for more context but tl;dr we need to be way more
aggressive about ignoring garbage generated by the anomaly detection
logic. This CL automates the ignorance of one kind of particularly
annoying garbage we've seen repeatedly.

I got hit with it on this week's on-call rotation, and I got hit
with it on my last rotation too.

A recurring source of pain is when, for ...whatever reason...
a device that was reporting measurments for some metric on the
waterfall is replaced by an entirely different physical device who
then proceeds to report the same metric but with a different bias
than the previous device. You end up with a "change" in the timeseries
but what it actually reflects is differences in physical device
performance, not differences in the code under test's performance.

This change takes the device ID reported by whatever uploaded a new
timeseries data point, and plumbs it through to the point in
AlertGroupWorkflow where we can compare it to the device ID that
reported the leading edge of the anomaly and unceremoniously drop
it in the trash with a logging.error statement.

Bug: chromium:1485723
Change-Id: I34347f1e67a7cb2159cec05c4b87fd6ed115730c