[chromeperf] use experiment_patch instead of patch when try job is created via the arrow button
The 'Test a patch' button was not updated to use 'base_patch' or 'experiment_patch' arguments to create a new try job. Instead, it still uses the 'patch' argument. The new try job can still run, but when end user tries to rerun that try job, the 'base_patch' and 'experiment_patch' cannot be populated as they don't exist.

This CL renamed the argument used in 'Test a patch' logic.

Bug: chromium:1483033
Change-Id: Ib62da7c1c9930a8271baba57dafa9f466c0ea80f