[Pinpoint] Don't fail when trying to add repository to Datastore
A Pinpoint job will raise an AssertionError and terminate the job early
if the repository in question already exists in Repository and the
system tries to add it in. This happens as the job is unravelling DEPS.

This change modifies the logic s.t. the name is returned early if the
object has matching URLs. Otherwise, the issue is logged as a warning.

We don't override the value in favor for existing data.

Bug: chromium:1491470
Change-Id: I47c8f0150250d8b9f9b1dbda86509a275e12b9ab