[chromeperf] fix the GetAllActiveAlertGroups to load ungrouped
Fixed the issue which causes the rollback: https://chromium-review.googlesource.com/c/catapult/+/4953065

The 'ungrouped' is supposed to be processed as the other regular groups. The change in https://chromium-review.googlesource.com/c/catapult/+/4836932 ignored the 'ungrouped' by defining the group_type. Here a extra step is added to load the missing group.

Change-Id: I6c39fb79f341ee1e8858aecc4a54164abe214b3b