Add logging to failing Anomaly calculation
See crbug/1510455, crbug/1510338.
Requests are erroring with 500. Internal logs show a error calculating
the abs delta for an anomaly (using before and after values). Both
these values are None, resulting in a TypeError.

This change adds a log message to be able to identify the Anomaly key
that's causing this error.

Bug: chromium:1510455, chromium:1510338
Change-Id: I4daa627a4f10541223b8085d657efc2da4c33a46