Add support for custom tag conflict checkers
Adds support for using custom tag conflict checkers in typ when running
SeriallyExecutedBrowserTests in Telemetry. The ability to use a custom
checker has existed for a long time, but only when using the expectation
parser code directly, not when typ is used to run tests.

crrev.com/c/5208211 is a use of this new functionality.

Bug: chromium:1502712
Change-Id: I487889ce424b37078562d90f389acde8b962c900