Shared Storage: Add inspector/tab method to wait for events
We add the ability to track events from Shared Storage by implementing
handling of "Storage.sharedStorageAccessed" DevTools protocol events
in `InspectorStorage` as well as to send requests to DevTools
Protocol for the method 'Storage.setSharedStorageTracking' (which can
enable or disable tracking based on the parameter).

We then use these capabilities to add, by way of `InspectorBackend`,
a `Tab` method `WaitForSharedStorageEvents()` that can wait for a
list of expected Shared Storage events in a given order.

We plan to use above method in perf benchmark tests for Shared Storage
that we are currently writing.

Information about the new Shared Storage API can be found at the
Shared Storage GitHub repo:
https://github.com/WICG/shared-storage/blob/main/README.md

Bug: chromium:1516507
Change-Id: I0439ee248696b26f4489de9b3be478ef689b6725