[chromeperf] update the apk name to fix pixel4 webview failure
On Pixel4 webview bots, Telemetry is looking for TrichromeLibraryGoogle.apk which is replaced by TrichromeLibraryGoogle6432.apk.

As pixel 4 webview is the only one browser using this setting, we will directly update the apk name. However, Pinpoint may still fail to run on the commits before this fix.

Pinpoint job to validate the fix:
https://pinpoint-dot-chromeperf.appspot.com/job/124f49f06e0000

Bug: chromium:1521416
Change-Id: I432c8a8c4f4a4ffb7fee096e98faff899b9662e4