Shared Storage: Add tab/inspector methods to get entries/metadata
This is a followup to https://crrev.com/c/5218592, after which I
realized that I likewise need methods for behnchmark tests to confirm
the current contents of shared storage for a origin, along with
metadata. Such methods already already exist in the DevTools Protocol,
so I hook up to them here.

Bug: chromium:1516507
Change-Id: I2f440e1d8b2af917331c7af3de6e9e5b0f4238c5