Shared Storage: Improve timeout handling in WaitForSharedStorageEvents
This is a followup to https://crrev.com/c/5218592 and
https://crrev.com/c/5238213.

In some cases, undispatched shared storage events are queued for the
websocket to receive, but they are not automatically being sent.

Note that `InspectorWebsocket.SyncRequest()` runs a loop calling
`InspectorWebsocket._Receive()` until the request's response is
received, so this has the side effect of flushing any queued
but undispatched protocol messages so that the websocket receives and
handles them.

Thus, if we timeout in waiting for shared storage events, we first
call `GetSharedStorageMetadata()`. This will give us additional debug
information should we actually raise the exception. More importantly,
however, since `GetSharedStorageMetadata()` calls
`InspectorWebsocket.SyncRequest()`, we exploit the observation above
to ensure delivery of any remaining events. At ths point, we may have
finally received the expected events and can return true. Otherwise,
we go ahead and raise the TimeoutException.

Bug: chromium:1516507
Change-Id: I2cdefb592eb1268da88712a03be8343507528172