Add OS Name to finder_options sent to the Benchmark
Benchmarks can be specified with either generic SUPPORTED_PLATFORMS such
as 'Desktop' or with OS level specificity such as 'Mac'. See
telemetry/telemetry/story/expectations.py for the full list.

Some benchmarks however may wish for some stories to only be ran on
certain OSes. Such as a benchmark supporting 'Desktop', but some stories
only on 'Mac'.

While telemetry/telemetry/benchmark.py confirms the SUPPORTED_PLATFORM
for the benchmark, the implementation of the benchmark builds the
story list via CreateStorySet.

This change adds OS Name to the `finder_options` so that the benchmark
can filter the stories based on OS as well.

This is to support Motionmark adding a Mac only variant for the
Skia-Graphite project. Which currently is only available on Mac, but
will expand to other OSes over time.

See: https://chromium-review.googlesource.com/c/chromium/src/+/5262797
Change-Id: I1c36f79503a2a8d4f836d038ac53f2c0dafd9cda