[chromeperf] Consider empty project names as chromium in grouping
When matching the project id in the grouping logic, the default value 'chromium' is not compared correctly with an empty string.

E.g., if the anomaly's subscription does not has monorail_project_id, and the group has the project_id as 'chromium', we will consider they have different project and thus fail to group the anomaly into the existing group.

Change-Id: I2df00387a54920135e6f56c5afd293aa6e4c8b96