[chromeperf] add alternatives for the *-NeedsAttention labels in monorail
Handling the use case of the two labels: Chromeperf-Sheriff-NeedsAttention and Chromeperf-Auto-NeedsAttention. They will be replaced by a hotlist Chromeperf-CulpritDetection-NeedsAttention.

For Chromeperf-Sheriff-NeedsAttention, now the text message will ask the endusers to add the issue to the hotlist instead.
For Chromeperf-Auto-NeedsAttention, a mapping is added so that the buganizer client to find the hotlist for it.

Bug: b/305074497
Change-Id: I3117ea91a6db057328d5fb99fc29d5aabc04bb91