telemetry: Remove unused arg, fix monochrome check
Previously the is_monochrome flag only accounted for Monochrome.apk and
monochrome_bundle. This CL expands it to include other variants.

Add logging to indicate which installation is for the support apk to
differentiate it from the main apk.

None of the implementations of FindSupportApks used the chrome_root
parameter, so this CL removes that paramenter from the call and all the
implementations.

Bug: chromium:325292888
Change-Id: Ic249461983fd00cd736e641ca126edc7f569b106