telemetry: Add --detailed option, improve logging
Add `--detailed` to the `list` command for //tools/perf/run_benchmark.
This allows listing all the stories with their names and tags from the
command line for a particular benchmark (e.g. rendering.mobile).

Reduce logging verbosity as each filtered story was logging three times,
once for each invocation of StoryFilter.ShouldSkip. Switch logging to be
turned on via a flag and only log once in story_runner.RunStorySet, this
way the rest of the ShouldSkip calls still work but do not add extra
verbosity via logging.

Refactor StoryFilter.ShouldSkip to be a bit easier to understand,
introducing intermediate boolean variables to highlight the two ways
that stories can have a valid reason to be disabled but still should run
in the current invocation.

Bug: 325292888
Change-Id: Ie5588b4c9b28de179b67883d7f786ea320e1dab1