Apply new thresholds (5% as minimal low threshold, 10% as the minimal high threshold) to the top two benchmarks to improve the autobisect culprit detect rate.
Based on the our dashboard, the benchmarks 'rendering.desktop' and 'system_health.common_desktop' caused 46.15% of the "failure" of autobisect jobs which wasn't able to find culprit in the last two weeks.

Apply the new thresholds to those top two benchmarks to align with the threshold (5%) of our verification workflow, which will improve our culprit detect rate.

Dashboard links:
1.
https://dashboards.corp.google.com/success_metrics_dash?p=rolling_window_size_gjnw22:1&p=rolling_window_size_z4creh:7&p=benchmark_hm9n2m:%22rendering.desktop%22,%22system_health.common_desktop%22&p=benchmark_z3z3gd:%22jetstream2%22#k6eumi
2.
https://pantheon.corp.google.com/monitoring/dashboards/builder/43a4b804-184d-413c-85d2-cbec0a15691e;duration=P14D?e=-13802955&mods=component_inspector&project=chromeperf&pli=1

Change-Id: Id09809534104680c1d8527ba77df6139c4f32ff7