Add logs to debug the bot id before anomaly.
There was waterfall bot change, which caused the different of testing result. We have made a CL (https://chromium-review.googlesource.com/c/catapult/+/4889182) to ignore such kind of anomalies.
But I think it has a bug, which might miss some of the bot changes.
This CL will be used to debug and verify that bug.
Once confirmed, will make another CL to fix that bug.


Change-Id: Ifaff968856eb6fc0b4b07f2d931b6d7009bb9fe7