Do not add a new ContentLength header when decoding responses
An archive can be edited with decoded responses using the following:
  go run src/httparchive.go edit --decode_response_body=True
    --full_path="..." in.wprgo out.wprgo
If a response in the archive does not have a ContentLength set, this
patch leaves it unset following the same pattern used in
`transformResponseBody`. A reason to not add ContentLength headers is
that it becomes one more field that must be manually edited if data
from the body is added/removed.

Change-Id: I9487c7a270b267eb58ff47268df5a4dfc3cbbc79