Add Telemetry flag to parse DUT hostnames from SWARMING_BOT_ID if running on ChromeOS Swarming.
The presence of '--fetch-cros-remote' flag tells Telemetry that it's running on ChromeOS Swarming. To decipher the DUT hostname, it must then parse the SWARMING_BOT_ID env var. We modify Pinpoint to use this flag when appropriate

Additionally, set ConnectTimeout to 30. For some reason, SSH'ing to a Skylab DUT will timeout after 5 seconds, so this needs to be increased.

Bug: b:313960680

Change-Id: Ia87382ca3005d95856fd0ce3c56f5e2a1b783049