[chromeperf] remove component from failed regression verification to min noise
Originally added `Speed>Dashboard` component to failed regression
verification jobs for the Browser Perf on-call sheriff to investigate.

However, in practice this generates too much noise and confusion for
the on-call to follow. Bugs that are closed with
`Regression-Verification-Failed` can also re-open due to the alert
grouping logic, generating emails for everyone on the team.

It is likely better to keep the `Regression-Verification-Failed` label
and periodically examine why these jobs are failing.

Bug: chromium:327686907
Change-Id: Idda4af9520275892570524f897ca9b70fcc22f1c