Make activity and action SDK-aware
Makes the activity and action components of browser intents SDK-aware so
that different versions of Android can use different `am start` commands
to start browsers. This is necessary due to intent filters on Android T+
which cause browser startup to fail on certain T+ devices when using the
old start command.

Bug: chromium:332350951
Change-Id: If08d0cb0a7047a1ac29816a43018096ff8cb346c