Add PowerShell timeout and retry
Adds a 10 second timeout and 3x retry to the PowerShell command used
to get the system type on Windows. It is possible for this command to
hang indefinitely when run on Windows ARM64 devices, but the issue
appears to be transient since other browsers before and after within the
same task are able to run the command without issue.

Bug: chromium:326122296
Change-Id: Icc59ae3923de561cafe0b808f931db419a2564fa