Add flaky device finding workaround
Adds --initial-find-device-attempts to the list of browser options,
which adjusts how many times devices are looked for when finding
browsers. The default value is 1, which results in no functional
changes.

This is used to work around Android devices going offline briefly,
which has been a low flake issue for a long time, but is being hit more
frequently on certain devices that were recently added.

Bug: chromium:333414298
Change-Id: I64382a483ef712133378f05f37c00ae1768b6007