telemetry: Symbolize minidumps with dump_syms
Update minidump_symbolizer.py to download dump_syms from the
BinaryManager symbolization. Previously it would fail to locate
dump_syms in the build output directory and fail symbolization.

Add assert to local_first_binary_manager.py to allow type checkers to
realize that it was not possible for cls._instance to be None.

Bug: chromium:332553182
Change-Id: I932b6602a0462cb2da524ef4414766d87fb5c1d1