[pinpoint] support post endpoint to inject a new job
we had originally planned for this endpoint to be in skia-bridge.
however, there posed two problems:
1. job state contains a map of change to attempts, where change
  is a namedtuple and contains commits, which is also a named
  tuple. Namedtuples as keys are not supported by the json library,
  requiring us to pickle the job state object before appending it to
  the job object.
2. pickling the job state means that the class definition must be
  importable and live in the same module as wehn the object is stored.
  doing this in skia-bridge imposes a problem because it's catapult
  that's unpickling the object.

this change thus introduces a post endpoint on /api/job that creates a
job object and pushes it to datastore. It does not make use of the
new() class method within job because that makes additional outbound
calls to generate some of its values, which is not needed for this
since the job has been run in skia.

this change is to support backwards compatibility from skia bisects
to catapult ui.

Bug: chromium:335543822
Change-Id: I3c4693e2a09c3e85cdf1f09a3245f2c96bee959c