[pinpoint] support uuid as id for Job models
Skia generates its workflows with UUIDs. Update the job model
to support uuid. It'll try to parse uuid (meaning that it doesn't
support any arbitrary string as an id, only uuid). If it fails,
it'll revert to int parsing as usual.

This change also updates the skia handler for writing Jobs into
datastore by appending a skia url to arguments so that it's
displayed on the side. It also stores the job as the uuid, meaning
the job id in pinpoint URL is also the same uuid.

Bug: chromium:340935348
Change-Id: I19c575dbbfeadbdf941406bd0f79e91859f77876