telemetry: Don't crawl all of //out/ when chromium-output-dir is passed
When --chromium-output-dir is passed, ideally telemetry only knows
to look for browser in that out dir. But for some reason it crawls
all of //out/ in that case looking for an out dir named the same as
the browser "type".

So instead this updates the desktop browser finder to only look in
the given out dir when it was specifically supplied.

Bug: chromium:342602628
Change-Id: I85cdcc53e5f27b2f60d2cd5eb830ed52da1c49e7