[bqexport/pinpoint] Ignore job export where ID != int
Skia based Pinpoint jobs run with a UUID as its ID (type string).
Making this compatible to Datastore worked because Datastore doesn't
require a schema. For BQ, the schema requires that the ID be int.

Although the UUID could be hexed and modified to int format to make
BQ work, it breaks consistency with the skia migration effort, because
we create dependencies back to Catapult services to accomplish this,
making it a band aid solution. Skia based Pinpoint jobs should export
its own data.

Thus, this change ignores all non int based Job IDs, meaning that
Skia based jobs that are written back to Datastore are ignored.

Bug: chromium:340935348
Change-Id: I0bb33fdb83693314dd30bd9e426f54722b40e190