[pinpoint] Reland support uuid as id for Job models
This is a reland of revert 2e1d4c5f8f18cdf8d6b9c3ac80c9bc8454197930

BQ export will ignore UUIDs that are not int based, meaning
that skia-based workflows are ignored.

Skia generates its workflows with UUIDs. Update the job model
to support uuid. It'll try to parse uuid (meaning that it doesn't
support any arbitrary string as an id, only uuid). If it fails,
it'll revert to int parsing as usual.

This change also updates the skia handler for writing Jobs into
datastore by appending a skia url to arguments so that it's
displayed on the side. It also stores the job as the uuid, meaning
the job id in pinpoint URL is also the same uuid.

Bug: chromium:340935348
Change-Id: I400a209003ed53cd7e9c78acddf573b3fcd2f731