[typ] Add URI encode/decode disable
Adds the option to disable URI encoding/decoding when parsing or
generating typ expectations. Some users of typ expectations such as GPU
do not assume that test names in expectation files are URLs, and thus
the encoding/decoding can cause issues.

The default behavior remains the same, so there should not be any
functional changes until users opt in.

Bug: chromium:346263468
Change-Id: I3f37f1bd2425e25f508279446d3a4660ab839a12