[pinpoint] omit swarming task cas output when task fails
When a skia bisection swarming task fails, MarshallToAttempt errors out
because it assumes that all swarming tasks will have an accompanying
cas isolate output. This change ensures to only append result args
when there is a isolate digest to consume.

Future changes might be needed to properly capture errors on swarming
tasks on the Pinpoint UI

Bug: chromium:335543316
Change-Id: I72b072d06129c0215d57095ad1ea3406a47a86ab