telemetry: Replace __del__ with atexit_with_log
Previously Forwarder.__del__ called self.Close, which seems fine in this
context, but its child class AndroidReverseForwarder.Close calls into
self._device.adb.ForwardRemove, which is also fine, it simply runs an
ADB command, but AdbWrapper._RunAdbCmd is decorated with
WithTimeoutAndConditionalRetries, which calls timeout_retry.Run, which
creates a TimeoutRetryThreadGroup (extending
reraiser_thread.ReraiserThreadGroup), and starts it, calling
ReraiserThreadGroup.StartAll which calls thread.start(). This then calls
into the python3 threading library and attempts to acquire locks in
order to start the thread. Herein is the problem, since __del__ is
called by the garbage collector, parts of python code is no longer
available during garbage collection, thus this acquisition sometimes
deadlocks here.

This deadlock doesn't always happen, it's hard to tell what initial
conditions results in this deadlock, but it can be consistently repro'ed
at high frequency (see linked bug) and the bot android-arm64-pgo
encounters it about 50/50 for every failing benchmark run. This deadlock
causes the benchmark to hang and the monitoring script to not detect
that it has finished, so this ends up taking the full 2 hours until the
grace period is finished and the monitoring script kills the underlying
benchmark script (which already completed, but was just hanging on
sys.exit).

This problem does not happen if atexit_with_log is used instead, since
this happens earlier in the shutdown process and not during garbage
collection. This also adds more clarity on what is being run at exit as
by passing -vvv the script outputs what atexit handlers are running,
whereas __del__ runs silently and does not log, causing this bug to be
hard to root-cause.

Thus, always prefer atexit_with_log over __del__ whenever possible.

Bug: chromium:41491803
Change-Id: Ic1780296daaa1eb62c506e8f178bb564658faa7f